refactor(developer): move message correctness test into shared unit
